backend/requirements.txt
# FastAPI & Server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
aiosqlite==0.20.0

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# AI (Google Gemini)
google-generativeai>=0.8.0

# Config
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Testing
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
